# majormate-server/docs/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "MajorMate API",
    "description": "Global Major-based Study Platform — Backend API\n\n## WebSocket / STOMP (실시간)\n\n연결 엔드포인트: `ws://localhost:8082/ws` (SockJS)\n\n### 구독 토픽\n| 토픽 | 설명 |\n|---|---|\n| `/topic/room.{roomId}` | 방 멤버 상태 브로드캐스트 |\n| `/topic/chat.{chatRoomId}` | 1:1 Q&A 채팅 메시지 수신 |\n| `/user/queue/errors` | 개인 에러 메시지 |\n\n### 전송 대상 (prefix: `/app`)\n| 대상 | Payload | 설명 |\n|---|---|---|\n| `/app/stopwatch.start` | `StartStopwatchRequest` | 스톱워치 시작 |\n| `/app/stopwatch.pause` | - | 일시정지 |\n| `/app/stopwatch.resume` | - | 재시작 |\n| `/app/stopwatch.end` | - | 종료 및 DB 저장 |\n| `/app/stopwatch.keyword` | `UpdateKeywordRequest` | 키워드 변경 |\n| `/app/chat.{chatRoomId}` | `ChatMessageRequest` | 채팅 메시지 전송 |\n\n### 브로드캐스트 메시지 타입 (`RoomBroadcastMessage.type`)\n`MEMBER_JOINED` · `MEMBER_LEFT` · `MEMBER_STARTED` · `MEMBER_PAUSED` · `MEMBER_RESUMED` · `MEMBER_ENDED` · `MEMBER_KEYWORD_UPDATED`",
    "version": "1.0.0"
  },
  "servers": [
    { "url": "http://localhost:8082", "description": "Local Dev" }
  ],
...
  "paths": {
...
    "/api/auth/me": {
      "get": {
        "tags": ["Auth"],
        "summary": "현재 인증된 사용자 정보 조회",
        "operationId": "getMe",
        "responses": {
          "200": {
            "description": "인증된 사용자",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/UserResponse" } } }
          },
          "401": { "description": "미인증" }
        }
      }
    },
    "/api/auth/google": {
      "post": {
        "tags": ["Auth"],
        "summary": "모바일 Google 로그인 (Access Token → JWT)",
        "operationId": "googleMobileAuth",
        "security": [],
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GoogleAuthRequest" } } }
        },
        "responses": {
          "200": {
            "description": "로그인 성공",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/GoogleAuthResponse" } } }
          },
          "401": { "description": "유효하지 않은 액세스 토큰 또는 이메일 미인증" }
        }
      }
    },
...
    "/api/qa/request": {
      "post": {
        "tags": ["QA"],
        "summary": "질문 요청 전송 (질문받기 허용 유저에게만 가능)",
        "operationId": "createQaRequest",
        "requestBody": {
          "required": true,
          "content": { "application/json": { "schema": { "$ref": "#/components/schemas/QaRequestCreateRequest" } } }
        },
        "responses": {
          "201": {
            "description": "질문 요청 생성",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/QaRequestResponse" } } }
          },
          "400": { "description": "대상 유저가 질문을 허용하지 않음" },
          "404": { "description": "대상 유저 없음" }
        }
      }
    },
    "/api/qa/{requestId}/accept": {
      "post": {
        "tags": ["QA"],
        "summary": "질문 요청 수락 → 1:1 채팅방 생성",
        "operationId": "acceptQaRequest",
        "parameters": [
          { "name": "requestId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "수락 완료. `chatRoomId` 포함",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/QaRequestResponse" } } }
          },
          "403": { "description": "응답 권한 없음" },
          "404": { "description": "요청 없음" }
        }
      }
    },
    "/api/qa/{requestId}/reject": {
      "post": {
        "tags": ["QA"],
        "summary": "질문 요청 거절",
        "operationId": "rejectQaRequest",
        "parameters": [
          { "name": "requestId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "거절 완료",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/QaRequestResponse" } } }
          },
          "403": { "description": "응답 권한 없음" },
          "404": { "description": "요청 없음" }
        }
      }
    },
    "/api/qa/chat/{chatRoomId}/history": {
      "get": {
        "tags": ["QA"],
        "summary": "채팅 내역 조회",
        "operationId": "getChatHistory",
        "parameters": [
          { "name": "chatRoomId", "in": "path", "required": true, "schema": { "type": "string", "format": "uuid" } }
        ],
        "responses": {
          "200": {
            "description": "채팅 메시지 목록 (시간순)",
            "content": {
              "application/json": {
                "schema": { "type": "array", "items": { "$ref": "#/components/schemas/ChatMessageResponse" } }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
      "GoogleAuthRequest": {
        "type": "object",
        "required": ["accessToken"],
        "properties": {
          "accessToken": { "type": "string", "description": "Google OAuth2 액세스 토큰 (userinfo API 호출용)" }
        }
      },
      "GoogleAuthResponse": {
        "type": "object",
        "properties": {
          "token":     { "type": "string", "description": "JWT (30일 유효, HS256)" },
          "isNewUser": { "type": "boolean", "description": "최초 가입 여부" }
        }
      },
...
      "UserResponse": {
        "type": "object",
        "properties": {
          "id":          { "type": "string", "format": "uuid" },
          "email":       { "type": "string", "format": "email" },
          "nickname":    { "type": "string" },
          "major":       { "type": "string" },
          "nationality": { "type": "string", "description": "ISO 3166-1 alpha-2 국가 코드" },
          "gender":      { "type": "string", "enum": ["MALE", "FEMALE", "OTHER", "PREFER_NOT_TO_SAY"] }
        }
      },
...
      "QaRequestCreateRequest": {
        "type": "object",
        "required": ["targetUserId"],
        "properties": {
          "targetUserId": { "type": "string", "format": "uuid", "description": "질문 대상 유저 ID (현재 공부 중 + 질문허용 상태여야 함)" },
          "message":      { "type": "string", "maxLength": 500, "nullable": true }
        }
      },
      "QaRequestResponse": {
        "type": "object",
        "properties": {
          "id":               { "type": "string", "format": "uuid" },
          "requesterId":      { "type": "string", "format": "uuid" },
          "requesterNickname": { "type": "string" },
          "targetId":         { "type": "string", "format": "uuid" },
          "targetNickname":   { "type": "string" },
          "message":          { "type": "string", "nullable": true },
          "status":           { "type": "string", "enum": ["PENDING", "ACCEPTED", "REJECTED"] },
          "chatRoomId":       { "type": "string", "format": "uuid", "nullable": true, "description": "수락 시 생성되는 1:1 채팅방 ID" },
          "createdAt":        { "type": "string", "format": "date-time" }
        }
      },
...
      "ChatMessageResponse": {
        "type": "object",
        "description": "STOMP `/topic/chat.{chatRoomId}` 브로드캐스트 메시지",
        "properties": {
          "id":             { "type": "string", "format": "uuid" },
          "chatRoomId":     { "type": "string", "format": "uuid" },
          "senderId":       { "type": "string", "format": "uuid" },
          "senderNickname": { "type": "string" },
          "content":        { "type": "string" },
          "createdAt":      { "type": "string", "format": "date-time" }
        }
      },
...
    },
    "securitySchemes": {
      "cookieAuth": {
        "type": "apiKey",
        "in": "cookie",
        "name": "JSESSIONID"
      },
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT"
      }
    }
  },
  "security": [{ "cookieAuth": [] }, { "bearerAuth": [] }]
}