# majormate-server/src/main/java/me/majormate/qa/dto/QaRequestCreateRequest.java
package me.majormate.qa.dto;

import jakarta.validation.constraints.NotNull;
import jakarta.validation.constraints.Size;

import java.util.UUID;

public record QaRequestCreateRequest(
        @NotNull UUID targetUserId,
        @Size(max = 500) String message
) {}